Used file-scoped namespaces and fixed typo 'Extentions'
