+++ backend/Squirrel.ConsoleApp/Program.cs
@@ -1,10 +1,9 @@
-﻿namespace Squirrel.ConsoleApp
+﻿namespace Squirrel.ConsoleApp;
+
+internal class Program
 {
-    internal class Program
+    static void Main(string[] args)
     {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello, World!");
-        }
+        Console.WriteLine("Hello, World!");
     }
 }
