+++ backend/Squirrel.Core/Squirrel.Core.BLL/MappingProfiles/SampleProfile.cs
@@ -1,16 +1,16 @@
 ﻿using AutoMapper;
 using Squirrel.Core.Common.DTO;
+using Squirrel.Core.Common.DTO.Sample;
 using Squirrel.Core.DAL.Entities;
 
-namespace Squirrel.Core.BLL.MappingProfiles
+namespace Squirrel.Core.BLL.MappingProfiles;
+
+public sealed class SampleProfile : Profile
 {
-    public sealed class SampleProfile : Profile
+    public SampleProfile()
     {
-        public SampleProfile()
-        {
-            CreateMap<Sample, SampleDto>();
-            CreateMap<SampleDto, Sample>();
-            CreateMap<NewSampleDto, Sample>();
-        }
+        CreateMap<Sample, SampleDto>();
+        CreateMap<SampleDto, Sample>();
+        CreateMap<NewSampleDto, Sample>();
     }
 }
