+++ backend/Squirrel.Core/Squirrel.Core.BLL/Services/Abstract/BaseService.cs
@@ -1,17 +1,16 @@
 ﻿using AutoMapper;
 using Squirrel.Core.DAL.Context;
 
-namespace Squirrel.Core.BLL.Services
-{
-    public abstract class BaseService
-    {
-        private protected readonly SquirrelCoreContext _context;
-        private protected readonly IMapper _mapper;
+namespace Squirrel.Core.BLL.Services.Abstract;
 
-        public BaseService(SquirrelCoreContext context, IMapper mapper)
-        {
-            _context = context;
-            _mapper = mapper;
-        }
+public abstract class BaseService
+{
+    private protected readonly SquirrelCoreContext _context;
+    private protected readonly IMapper _mapper;
+
+    public BaseService(SquirrelCoreContext context, IMapper mapper)
+    {
+        _context = context;
+        _mapper = mapper;
     }
 }
