+++ backend/Squirrel.Core/Squirrel.Core.Common/DTO/Sample/NewSampleDto.cs
@@ -1,9 +1,8 @@
-﻿namespace Squirrel.Core.Common.DTO
+﻿namespace Squirrel.Core.Common.DTO.Sample;
+
+public class NewSampleDto
 {
-    public class NewSampleDto
-    {
-        public string Title { get; set; } = string.Empty;
-        public string? Body { get; set; }
-        public int CreatedBy { get; set; }
-    }
+    public string Title { get; set; } = string.Empty;
+    public string? Body { get; set; }
+    public int CreatedBy { get; set; }
 }
