+++ backend/Squirrel.Core/Squirrel.Core.Common/DTO/Sample/SampleDto.cs
@@ -1,11 +1,10 @@
-﻿namespace Squirrel.Core.Common.DTO
+﻿namespace Squirrel.Core.Common.DTO.Sample;
+
+public class SampleDto
 {
-    public class SampleDto
-    {
-        public long Id { get; set; }
-        public string Title { get; set; } = string.Empty;
-        public string? Body { get; set; }
-        public int CreatedBy { get; set; }
-        public DateTime CreatedAt { get; set; }
-    }
+    public long Id { get; set; }
+    public string Title { get; set; } = string.Empty;
+    public string? Body { get; set; }
+    public int CreatedBy { get; set; }
+    public DateTime CreatedAt { get; set; }
 }
