+++ backend/Squirrel.Core/Squirrel.Core.DAL/Context/EntityConfigurations/SampleConfig.cs
@@ -1,17 +1,16 @@
-﻿using Squirrel.Core.DAL.Entities;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using Squirrel.Core.DAL.Entities;
 
-namespace Squirrel.Core.DAL.Context
+namespace Squirrel.Core.DAL.Context.EntityConfigurations;
+
+public class SampleConfig : IEntityTypeConfiguration<Sample>
 {
-    public class SampleConfig : IEntityTypeConfiguration<Sample>
+    public void Configure(EntityTypeBuilder<Sample> builder)
     {
-        public void Configure(EntityTypeBuilder<Sample> builder)
-        {
-            builder.Property(e => e.Title)
-             .IsRequired()
-             .HasMaxLength(50);
+        builder.Property(e => e.Title)
+               .IsRequired()
+               .HasMaxLength(50);
 
-        }
     }
 }
