+++ backend/Squirrel.Core/Squirrel.Core.DAL/Context/SquirrelCoreContext.cs
@@ -1,20 +1,19 @@
 ﻿using Squirrel.Core.DAL.Entities;
 using Microsoft.EntityFrameworkCore;
 
-namespace Squirrel.Core.DAL.Context
+namespace Squirrel.Core.DAL.Context;
+
+public class SquirrelCoreContext : DbContext
 {
-    public class SquirrelCoreContext : DbContext
+    public DbSet<Sample> Samples => Set<Sample>();
+
+    public SquirrelCoreContext(DbContextOptions<SquirrelCoreContext> options) : base(options)
     {
-        public DbSet<Sample> Samples => Set<Sample>();
+    }
 
-        public SquirrelCoreContext(DbContextOptions<SquirrelCoreContext> options) : base(options)
-        {
-        }
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            modelBuilder.Configure();
-            modelBuilder.Seed();
-        }
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        modelBuilder.Configure();
+        modelBuilder.Seed();
     }
 }
