+++ backend/Squirrel.Core/Squirrel.Core.DAL/Entities/Common/AuditEntity/AuditEntity.cs
@@ -1,8 +1,7 @@
-﻿namespace Squirrel.Core.DAL.Entities
+﻿namespace Squirrel.Core.DAL.Entities.Common.AuditEntity;
+
+public abstract class AuditEntity<T> : Entity<T> where T : struct
 {
-    public abstract class AuditEntity<T> : Entity<T> where T : struct
-    {
-        public T? CreatedBy { get; set; }
-        public DateTime CreatedAt { get; set; }
-    }
+    public T? CreatedBy { get; set; }
+    public DateTime CreatedAt { get; set; }
 }
