+++ backend/Squirrel.Core/Squirrel.Core.DAL/Entities/Common/Entity.cs
@@ -1,7 +1,6 @@
-﻿namespace Squirrel.Core.DAL.Entities
+﻿namespace Squirrel.Core.DAL.Entities.Common;
+
+public abstract class Entity<T> where T : struct
 {
-    public abstract class Entity<T> where T : struct
-    {
-        public T Id { get; set; }
-    }
+    public T Id { get; set; }
 }
