+++ backend/Squirrel.Core/Squirrel.Core.DAL/Entities/Sample.cs
@@ -1,8 +1,9 @@
-﻿namespace Squirrel.Core.DAL.Entities
+﻿using Squirrel.Core.DAL.Entities.Common.AuditEntity;
+
+namespace Squirrel.Core.DAL.Entities;
+
+public class Sample : AuditEntity<long>
 {
-    public class Sample : AuditEntity<long>
-    {
-        public string Title { get; set; } = string.Empty;
-        public string? Body { get; set; }
-    }
+    public string Title { get; set; } = string.Empty;
+    public string? Body { get; set; }
 }
