+++ backend/Squirrel.Core/Squirrel.Core.WebAPI/Extensions/ApplicationBuilderExtensions.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Squirrel.Core.DAL.Context;
+
+namespace Squirrel.Core.WebAPI.Extensions;
+
+public static class ApplicationBuilderExtensions
+{
+    public static void UseSquirrelCoreContext(this IApplicationBuilder app)
+    {
+        using var scope = app.ApplicationServices.GetService<IServiceScopeFactory>()?.CreateScope();
+        using var context = scope?.ServiceProvider.GetRequiredService<SquirrelCoreContext>();
+        context?.Database.Migrate();
+    }
+}
