--- backend/Squirrel.Core/Squirrel.Core.WebAPI/Extentions/ApplicationBuilderExtensions.cs
@@ -1,15 +1,0 @@
-﻿using Squirrel.Core.DAL.Context;
-using Microsoft.EntityFrameworkCore;
-
-namespace Squirrel.Core.WebAPI.Extentions
-{
-    public static class ApplicationBuilderExtensions
-    {
-        public static void UseSquirrelCoreContext(this IApplicationBuilder app)
-        {
-            using var scope = app.ApplicationServices.GetService<IServiceScopeFactory>()?.CreateScope();
-            using var context = scope?.ServiceProvider.GetRequiredService<SquirrelCoreContext>();
-            context?.Database.Migrate();
-        }
-    }
-}
