+++ backend/Squirrel.Core/Squirrel.Core.WebAPI/Validators/Sample/NewSampleDtoValidator.cs
@@ -1,14 +1,13 @@
-﻿using Squirrel.Core.Common.DTO;
-using FluentValidation;
+﻿using FluentValidation;
+using Squirrel.Core.Common.DTO.Sample;
 
-namespace Squirrel.Core.WebAPI.Validators
+namespace Squirrel.Core.WebAPI.Validators.Sample;
+
+public class NewSampleDtoValidator : AbstractValidator<SampleDto>
 {
-    public class NewSampleDtoValidator : AbstractValidator<SampleDto>
+    public NewSampleDtoValidator()
     {
-        public NewSampleDtoValidator()
-        {
-            RuleFor(u => u.Title).NotNull().MaximumLength(50);
-            RuleFor(u => u.CreatedBy).NotEmpty();
-        }
+        RuleFor(u => u.Title).NotNull().MaximumLength(50);
+        RuleFor(u => u.CreatedBy).NotEmpty();
     }
 }
