+++ backend/Squirrel.Notifier/Squirrel.Notifier.WebAPI/Controllers/BroadcastController.cs
@@ -1,23 +1,22 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.SignalR;
-using Squirrel.Notifier.Hubs;
-using Squirrel.Notifier.Hubs.Interfaces;
+using Squirrel.Notifier.WebAPI.Hubs;
+using Squirrel.Notifier.WebAPI.Hubs.Interfaces;
 
-namespace Squirrel.Notifier.Controllers
+namespace Squirrel.Notifier.WebAPI.Controllers;
+
+[ApiController]
+[Route("[controller]")]
+public class BroadcastController : ControllerBase
 {
-    [ApiController]
-    [Route("[controller]")]
-    public class BroadcastController : ControllerBase
+    private readonly IHubContext<BroadcastHub, IBroadcastHubClient> _hubContext;
+
+    public BroadcastController(IHubContext<BroadcastHub, IBroadcastHubClient> hubContext)
     {
-        private readonly IHubContext<BroadcastHub, IBroadcastHubClient> _hubContext;
-
-        public BroadcastController(IHubContext<BroadcastHub, IBroadcastHubClient> hubContext)
-        {
-            _hubContext = hubContext;
-        }
-
-        [HttpPost]
-        public Task BroadcastMessage(string message)
-            => _hubContext.Clients.All.BroadcastMessage(message);
+        _hubContext = hubContext;
     }
+
+    [HttpPost]
+    public Task BroadcastMessage(string message)
+        => _hubContext.Clients.All.BroadcastMessage(message);
 }
